backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
python-dotenv==1.0.1
pytest==8.3.3
requests==2.32.3